refactor(api): extract task patches + artifacts handlers
handler_tasks.go was 2331 LOC, 32 funcs across multiple concerns.
Pull the patches + artifacts cluster out into a sibling file so
each handler file holds one thing.

  handler_tasks.go: 2331 → 2024 LOC
  handler_tasks_patches.go: new, 324 LOC

What moves to handler_tasks_patches.go

  Handlers (8)
  - HandleTaskArtifacts
  - HandleTaskRunArtifacts
  - HandleTaskRunArtifact
  - HandleTaskRunPatches
  - HandleTaskRunPatch
  - HandleRevertTaskRunPatch
  - HandleApplyTaskRunPatch
  - (loadTaskRunPatch helper, called only from the apply / revert
    handlers above)

  Helpers used only by the patch apply path
  - verifyPatchApplyPrecondition
  - pathWithinRoot

What stays in handler_tasks.go

  CRUD + lifecycle + stream + events + run-step handlers and the
  shared loaders (loadAuthorizedTask, loadAuthorizedTaskRun,
  decodeTaskRunEventData, buildTaskRunStreamState, parseAfterSequence,
  buildTaskItem / buildTaskActivityItems and their helpers,
  newTaskID, newOpaqueTaskResourceID, writeSSEHeaders,
  normalizeApprovalDecision).

The receiver type stays *Handler everywhere; no public API
changes; the mux registrations in server.go are unchanged.
Imports adjusted: handler_tasks.go drops the now-unused `os`
import; handler_tasks_patches.go takes its own
`context / errors / fmt / log/slog / net/http / os / path/filepath
/ strings / time` + `internal/taskstate / internal/telemetry /
pkg/types`.

Verification

  go build ./...                — clean
  go vet ./...                  — clean
  go test ./... -race -count=1  — 1779 / 1779 pass
